Move mixins to the separate folder
