+++ src/ggrc-client/js/models/mixins/base-notifications.js
@@ -1,0 +1,11 @@
+/*
+    Copyright (C) 2018 Google Inc.
+    Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+*/
+
+import Mixin from './mixin';
+
+export default Mixin('base-notifications', {
+  send_by_default: true,
+  recipients: 'Admin,Primary Contacts,Secondary Contacts',
+});
