+++ src/ggrc-client/js/models/mixins/index.js
@@ -1,0 +1,24 @@
+/*
+  Copyright (C) 2018 Google Inc.
+  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+*/
+
+import './access-control-list';
+import './assertions-categories';
+import './assessment-issue-tracker';
+import './audit-issue-tracker';
+import './auto-status-changeable';
+import './base-notifications';
+import './ca-update';
+import './contactable';
+import './in-scope-objects-preload';
+import './in-scope-objects';
+import './is-overdue';
+import './mapping-limit-issue';
+import './mapping-limit';
+import './mixin';
+import './proposable';
+import './refetch-hash';
+import './related-assessments-loader';
+import './timeboxed';
+import './unique-title';
