+++ src/ggrc-client/js/models/mixins/mapping-limit-issue.js
@@ -1,0 +1,15 @@
+/*
+    Copyright (C) 2018 Google Inc.
+    Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+*/
+
+import Mixin from './mixin';
+
+export default Mixin('mapping-limit-issue', {
+  getAllowedMappings() {
+    return _.union(
+      GGRC.config.snapshotable_objects,
+      ['Issue', 'Program', 'Project', 'TaskGroup', 'Document']
+    );
+  },
+}, {});
